+++ src/admin/role.js
@@ -1,10 +1,11 @@
-﻿export const roleInit = async (client, commons) => {
+﻿import { COMMONS } from "../commons.js";
+
+export const roleInit = async (client) => {
   console.log("role init");
   // Client init and check reactions on role message
   const server =
-    process.env.DEBUG === "yes"
-      ? commons.find(({ name }) => name === "test")
-      : commons.find(({ name }) => name === "prod");
+    process.env.DEBUG === "yes" ? COMMONS.getTest() : COMMONS.getProd();
+
   const rolesJson = Object.values(server.roles); //get all the roles we are working with - format : [color, {roleId:, name:}]
 
   //check if the message has all Ewibot reactions