+++ src/bot.js
@@ -1,3 +1,6 @@
+import dotenv from "dotenv";
+dotenv.config();
+
 import dayjs from "dayjs";
 import RelativeTime from "dayjs/plugin/relativeTime.js";
 import "dayjs/locale/fr.js";
@@ -52,8 +55,7 @@ import {
 import { initAdminLogClearing } from "./admin/utils.js";
 
 // jsons import
-import { readFileSync } from "fs";
-const commons = JSON.parse(readFileSync("static/commons.json"));
+import { COMMONS } from "./commons.js";
 
 // commands imports
 import { wishBirthday } from "./commands/birthday.js";
@@ -90,9 +92,9 @@ const frequency = 24 * 60 * 60 * 1000; // 24 hours in ms
 
 setTimeout(async () => {
   // init birthday check
-  const server = commons.find(({ name }) =>
-    process.env.DEBUG === "yes" ? name === "test" : name === "prod"
-  );
+  const server =
+    process.env.DEBUG === "yes" ? COMMONS.getTest() : COMMONS.getProd();
+
   const channel = await client.channels.fetch(server.randomfloodChannelId);
 
   console.log("hello, timeoutBirthday");
@@ -147,9 +149,7 @@ const onMessageHandler = async (message) => {
   if (channel.type === "DM") {
     onPrivateMessage(message, client);
   } else {
-    const currentServer = commons.find(
-      ({ guildId }) => guildId === channel.guild.id
-    );
+    const currentServer = COMMONS.fetchGuildId(channel.guildId);
     onPublicMessage(message, client, currentServer, self);
   }
 };
@@ -158,15 +158,15 @@ const onMessageHandler = async (message) => {
 client.once("ready", async () => {
   // Bot init
   console.log("I am ready!");
-  roleInit(client, commons); //role handler init
+  roleInit(client); //role handler init
 
   //Ewibot activity
   setActivity(client);
   updateActivity(client);
 
-  const server = commons.find(({ name }) =>
-    process.env.DEBUG === "yes" ? name === "test" : name === "prod"
-  );
+  //slash commands
+  const server =
+    process.env.DEBUG === "yes" ? COMMONS.getTest() : COMMONS.getProd();
   const guildId = server.guildId;
   slashCommandsInit(self, guildId, client); //commands submit to API
 
@@ -177,19 +177,18 @@ client.once("ready", async () => {
   client.twitter.isSending = false;
   initTwitterLoop(client);
 
-  // LOGS
-
+  //LOGS
   const tomorrow2Am = dayjs()
     .add(1, "day")
     .hour(2)
     .minute(0)
     .second(0)
     .millisecond(0); //tomorrow @ 2am
-
   const timeTo2Am = tomorrow2Am.diff(dayjs()); //10000; //waiting time in ms
   initAdminLogClearing(client, timeTo2Am); //adminLogs clearing init
 
-  setGiftTimeoutLoop(client, commons); //gift timeout loop init
+  //gift
+  setGiftTimeoutLoop(client); //gift timeout loop init
 });
 // Create an event listener for messages