+++ changelog.md
@@ -1,10 +1,16 @@
 ﻿## 5.0.0
+- New: `commons.js` file - `Class Commons` with same behaviour as `Class Personality`
 - Remove: `$commands` (birthday, concrete, help, ignoreUser, ignoreChannel, ping, reminder, roll)
 - Remove: $ access to remaining `$commands` (leadApo, pronouns, spotify)
 - Remove: `$commands` helpers (checkIsOnThread, isCommand)
+- Update: files now use `COMMONS` object instead of commons.json parse 
+- Update: personality organisation 
+	- explode personalities.json file into 3 files
+	- regroup files into personalities folder
 
 ## 4.8.0
 - Update: `reverse-translator` now can translate frequent-logs
+- Update: add gift button date check + change gift command release date 
 
 ## 4.7.0
 - Feat: New Year personality